trying to pass the correct id number

+++ web/week05/confirm_delete.php
@@ -1,6 +1,6 @@
 <?php 
   $name = htmlspecialchars($_POST['meal']);
-  $id   = (int)htmlspecialchars($_POST['id']);
+  $id   = htmlspecialchars($_POST['id']);
 
   session_start();
